backend/api/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from api import models


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('url', 'username', 'email', 'groups')


class ClientSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = models.Client
        fields = ('url', 'name', 'adress', 'company')
